we do a little refactoring :trollface:
